% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_benefit_score.R
\name{convert_to_benefit_score}
\alias{convert_to_benefit_score}
\title{Converts raw scores to project benefit scores}
\usage{
convert_to_benefit_score(data)
}
\arguments{
\item{data}{a dataframe with 29 columns. The first is the project name. The next 28 are the metrics. With a seperate column for metric 6a and 6b. Each proposals is a row.}
}
\description{
\code{authors_clean} This function takes the output from `references_read()` and cleans the author information.
}
