% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_to_utility.R
\docType{data}
\name{scores_to_utility}
\alias{scores_to_utility}
\title{Scores to Utility Conversion Table}
\format{a table of how to convert between scores and utilities
\describe{
  \item{score_to_utility}{a dataframe}
  }}
\usage{
scores_to_utility
}
\description{
#'
}
\keyword{datasets}
