% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\docType{data}
\name{weights}
\alias{weights}
\title{Weights for each performance metric}
\format{a table of performance metric weights
\describe{
  \item{weights}{a dataframe}
  }}
\usage{
weights
}
\description{
#'
}
\keyword{datasets}
